Linting fixes

+++ 3DRenderer/utils/tests/test_clip_triangles.cpp
@@ -22,7 +22,9 @@ static Varying makeVarying(float x, float y, float z, int r = 255,
                            int g = 255, int b = 255) {
     Varying v;
     v.position = Eigen::Vector4f(x, y, z, 1.0f);
-    v.color = Eigen::Vector3f(r, g, b);
+    v.color =
+        Eigen::Vector3f(static_cast<float>(r), static_cast<float>(g),
+                        static_cast<float>(b));
     return v;
 }
